Refactor: Improve code quality, readability, and DRYness
This commit refactors the GoalTracker application to improve code quality, readability, and adherence to the DRY (Don't Repeat Yourself) principle.

The main changes include:

- **Componentization**: Broke down the monolithic `DashboardScreen` and `GoalSettingScreen` into smaller, reusable components (`GoalCard` and `GoalForm`). This improves maintainability and readability.
- **State Management**: Simplified the state management in `GoalSettingScreen` by using a single state object for each goal, reducing complexity and potential for errors.
- **Code Organization**: Centralized type definitions in `types/goals.ts` and moved constants to their own files (`constants/units.ts`).
- **Styling**: Refactored the styling to avoid creating stylesheets on every render, improving performance. Replaced hardcoded colors with values from `constants/Colors.ts`.
- **Type Safety**: Improved type safety in the database layer by using defined types instead of `any`.
