+++ GoalTracker/app/dashboard.tsx
@@ -2,7 +2,6 @@ import { useRouter } from 'expo-router';
 import React, { useEffect, useState } from 'react';
 import {
     Alert,
-    Dimensions,
     Image,
     ScrollView,
     StyleSheet,
@@ -11,29 +10,15 @@ import {
     useColorScheme,
     View,
 } from 'react-native';
-import { LineChart } from 'react-native-chart-kit';
+import GoalCard from '../components/GoalCard';
 import Colors from '../constants/Colors';
 import { deleteGoal, getGoalProgress, getGoals } from '../database';
-import { Goal } from '../types/goals';
-
-interface GoalProgress {
-  id: number;
-  goalId: number;
-  currentValue: number;
-  date: string;
-}
-
-interface GoalWithProgress extends Goal {
-  progress: GoalProgress[];
-  latestProgress?: number;
-  progressPercentage: number;
-}
+import { GoalWithProgress } from '../types/goals';
 
 const DashboardScreen = () => {
   const colorScheme = useColorScheme();
   const styles = getStyles(colorScheme);
   const router = useRouter();
-  const screenWidth = Dimensions.get('window').width;
   
   const [goalsWithProgress, setGoalsWithProgress] = useState<GoalWithProgress[]>([]);
   const [isLoading, setIsLoading] = useState(true);
@@ -96,114 +81,6 @@ const DashboardScreen = () => {
     );
   };
 
-  const renderProgressChart = (goal: GoalWithProgress) => {
-    if (goal.progress.length === 0) {
-      return (
-        <View style={styles.noDataContainer}>
-          <Text style={styles.noDataText}>No progress data yet</Text>
-          <Text style={styles.noDataSubtext}>Start tracking your progress!</Text>
-        </View>
-      );
-    }
-
-    const chartData = {
-      labels: goal.progress.slice(-7).map(p => {
-        const date = new Date(p.date);
-        return `${date.getMonth() + 1}/${date.getDate()}`;
-      }),
-      datasets: [
-        {
-          data: goal.progress.slice(-7).map(p => p.currentValue),
-          color: (opacity = 1) => `rgba(0, 123, 255, ${opacity})`,
-          strokeWidth: 2,
-        },
-      ],
-    };
-
-    return (
-      <LineChart
-        data={chartData}
-        width={screenWidth - 40}
-        height={180}
-        chartConfig={{
-          backgroundColor: styles.chartBackground.backgroundColor,
-          backgroundGradientFrom: styles.chartBackground.backgroundColor,
-          backgroundGradientTo: styles.chartBackground.backgroundColor,
-          decimalPlaces: 1,
-          color: (opacity = 1) => `rgba(0, 0, 0, ${opacity})`,
-          labelColor: (opacity = 1) => `rgba(0, 0, 0, ${opacity})`,
-          style: {
-            borderRadius: 16,
-          },
-          propsForDots: {
-            r: '4',
-            strokeWidth: '2',
-            stroke: '#007BFF',
-          },
-        }}
-        bezier
-        style={styles.chart}
-      />
-    );
-  };
-
-  const renderProgressBar = (goal: GoalWithProgress) => {
-    return (
-      <View style={styles.progressBarContainer}>
-        <View style={styles.progressBarBackground}>
-          <View 
-            style={[
-              styles.progressBarFill, 
-              { width: `${goal.progressPercentage}%` }
-            ]} 
-          />
-        </View>
-        <Text style={styles.progressText}>
-          {goal.latestProgress?.toFixed(1) || '0'} / {goal.targetNumber} {goal.targetUnit}
-        </Text>
-        <Text style={styles.percentageText}>
-          {goal.progressPercentage.toFixed(1)}%
-        </Text>
-      </View>
-    );
-  };
-
-  const renderGoalCard = (goal: GoalWithProgress) => (
-    <TouchableOpacity 
-      key={goal.id} 
-      style={styles.goalCard}
-      onPress={() => router.push(`/goal-detail?goalId=${goal.id}`)}
-    >
-      <View style={styles.goalHeader}>
-        <View style={styles.goalHeaderLeft}>
-          <Text style={styles.goalName}>{goal.name}</Text>
-          <Text style={styles.goalStatus}>
-            {goal.completed ? 'Completed' : 'In Progress'}
-          </Text>
-        </View>
-        <TouchableOpacity
-          style={styles.deleteButton}
-          onPress={(e) => {
-            e.stopPropagation();
-            handleDeleteGoal(goal);
-          }}
-        >
-          <Text style={styles.deleteButtonText}>Delete</Text>
-        </TouchableOpacity>
-      </View>
-      
-      <Text style={styles.goalDescription}>{goal.description}</Text>
-      <Text style={styles.targetDate}>Target Date: {goal.targetDate}</Text>
-      
-      {renderProgressBar(goal)}
-      
-      <View style={styles.chartContainer}>
-        <Text style={styles.chartTitle}>Progress Over Time</Text>
-        {renderProgressChart(goal)}
-      </View>
-    </TouchableOpacity>
-  );
-
   if (isLoading) {
     return (
       <View style={styles.container}>
@@ -260,7 +137,9 @@ const DashboardScreen = () => {
 
           <View style={styles.goalsSection}>
             <Text style={styles.sectionTitle}>Your Goals</Text>
-            {goalsWithProgress.map(renderGoalCard)}
+            {goalsWithProgress.map((goal) => (
+              <GoalCard key={goal.id} goal={goal} onDelete={handleDeleteGoal} />
+            ))}
           </View>
 
           <View style={styles.actionButtons}>
@@ -372,106 +251,6 @@ const getStyles = (colorScheme: 'light' | 'dark' | null | undefined) => {
       color: colors.text,
       marginBottom: 15,
     },
-    goalCard: {
-      backgroundColor: colors.card,
-      padding: 20,
-      borderRadius: 15,
-      marginBottom: 20,
-      borderWidth: 1,
-      borderColor: colors.border,
-    },
-    goalHeader: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      alignItems: 'center',
-      marginBottom: 10,
-    },
-    goalHeaderLeft: {
-      flex: 1,
-    },
-    goalName: {
-      fontSize: 18,
-      fontWeight: 'bold',
-      color: colors.text,
-    },
-    goalStatus: {
-      fontSize: 12,
-      color: '#28a745',
-      fontWeight: '500',
-    },
-    goalDescription: {
-      fontSize: 14,
-      color: colors.text,
-      marginBottom: 8,
-      opacity: 0.8,
-    },
-    targetDate: {
-      fontSize: 12,
-      color: colors.text,
-      marginBottom: 15,
-      opacity: 0.7,
-    },
-    progressBarContainer: {
-      marginBottom: 20,
-    },
-    progressBarBackground: {
-      height: 8,
-      backgroundColor: colors.border,
-      borderRadius: 4,
-      marginBottom: 8,
-    },
-    progressBarFill: {
-      height: '100%',
-      backgroundColor: '#007BFF',
-      borderRadius: 4,
-    },
-    progressText: {
-      fontSize: 14,
-      color: colors.text,
-      fontWeight: '500',
-    },
-    percentageText: {
-      fontSize: 12,
-      color: colors.text,
-      opacity: 0.7,
-      marginTop: 2,
-    },
-    chartContainer: {
-      marginTop: 10,
-    },
-    chartTitle: {
-      fontSize: 16,
-      fontWeight: '600',
-      color: colors.text,
-      marginBottom: 10,
-    },
-    chart: {
-      marginVertical: 8,
-      borderRadius: 16,
-    },
-    chartBackground: {
-      backgroundColor: colors.background,
-    },
-    noDataContainer: {
-      height: 180,
-      justifyContent: 'center',
-      alignItems: 'center',
-      backgroundColor: colors.background,
-      borderRadius: 16,
-      borderWidth: 1,
-      borderColor: colors.border,
-    },
-    noDataText: {
-      fontSize: 16,
-      color: colors.text,
-      opacity: 0.7,
-    },
-    noDataSubtext: {
-      fontSize: 12,
-      color: colors.text,
-      opacity: 0.5,
-      marginTop: 5,
-    },
     actionButtons: {
       flexDirection: 'row',
       paddingHorizontal: 20,
@@ -479,7 +258,7 @@ const getStyles = (colorScheme: 'light' | 'dark' | null | undefined) => {
       gap: 10,
     },
     primaryButton: {
-      backgroundColor: '#007BFF',
+      backgroundColor: Colors.light.tint,
       paddingVertical: 15,
       paddingHorizontal: 30,
       borderRadius: 8,
@@ -503,18 +282,6 @@ const getStyles = (colorScheme: 'light' | 'dark' | null | undefined) => {
     secondaryButtonText: {
       color: colors.text,
       fontSize: 14,
-      fontWeight: '500',
-    },
-    deleteButton: {
-      backgroundColor: '#dc3545',
-      paddingHorizontal: 12,
-      paddingVertical: 6,
-      borderRadius: 6,
-      marginLeft: 10,
-    },
-    deleteButtonText: {
-      color: '#FFFFFF',
-      fontSize: 12,
       fontWeight: '500',
     },
   });
