+++ GoalTracker/constants/units.ts
@@ -1,0 +1,1 @@
+export const validUnits = ['km', 'miles', 'lbs', 'kg'];
